Move the JshellSecurityManager to the sandbox package
